Create session and guards
Note:
I have not finished it at the commit.
I just implement it first. :)

In the previous commit, I installed a package called 'express-session'.
This package is used to store and manage the user state across requests.
It usually goes with the middleware 'session' provided by the package 'passport'.
passport manages the user state with the session for us.

Next, I created a custom Guard called 'LoginGuard', which inherits AuthGuard('local').
Besides validating whether the user , I grab the request object from the context and
call super.logIn with the request object as its first parameter.
What is does is asking passport to record the user's login state. (Make the user 'logged in').

After finishing writing LoginGuard, I wrote another Guard called 'AuthenticatedGuard'.
This guard is used for routes to check whether the user is logged in.
If the user has log in successfully, the user's request will be handed to the route handler.
Otherwise, passport will respond with status code 401.
This class inherits the interface 'CanActivate'.
It basically just return the result of passport authentication. (request.isAuthentication)

So far, we have two Guard: 'LoginGuard' and 'AuthenticatedGuard'.
One is for user login. The other is for authentication.

There is still a thing I haven't metioned:
"How passport memorize our users' state?"
Passport needs to know how it stores our users' state.
Since Nest.js integrates with passport, it provides the class PassportSerializer for us to define
how it stores the state.
So I created a class called 'SessionSerializer', which inherits PassportSerializer.
Next, I overwrote two functions: 'serializeUser' and 'deserializeUser'.
The former is used to give data that will be stored in the session to passport.
The latter is used to extract the user data from the database by the data stored in the session.
By the defition, our users' data should be stored in the session successfully.

That's all, it should work when apply them on the route.